Fix error message if response code was not 200
Currently prcat does not handle the situation in where it read a part
of the response which comes after the headers. If the return code is
not 200 OK, the chances are high that it reads past them. Moving the
code block to part it matters (when code is 200) fixes this issue.